Internal: moved cmd to a separate class
